requirements.txt
torch
numpy
opencv-python
shapely
Pillow